change also to minitest gem in test helper file
Hi,

A previous pull request changed to the minitest gem in the Rakefile. There is a leftover in the test helper file, if one wants to call the tests directly without rake. This change reflects the move to minitest gem also in the helper file used for the tests.

Cheers

Cédric